facebook/requirements.txt
amqp==2.6.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
black==23.1.0
boto3==1.20.42
botocore==1.23.54
celery==4.4.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.11
django-allauth==0.52.0
django-celery-beat==2.0.0
django-cors-headers==3.11.0
django-timezone-field==4.2.3
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
flake8==6.0.0
flower==0.9.7
gspread==3.1.0
gunicorn==20.1.0
httplib2==0.21.0
humanize==4.4.0
idna==3.4
jmespath==0.10.0
kombu==4.6.11
mccabe==0.7.0
mypy-extensions==0.4.3
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.0
pathspec==0.10.3
platformdirs==2.6.2
prometheus-client==0.8.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.7.1
redis==4.1.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tornado==6.2
typing_extensions==4.1.0
tzdata==2022.7
urllib3==1.26.11
vine==1.3.0
wcwidth==0.2.5
wrapt==1.14.1
